nginx* fix upload gateway timeout
When uploading large files, users would sometimes encounter a gateway
timeout due to the fact that nginx was dropping the connection to the
backend after a set amount of time and indicating that the upload may
have failed, although the backend would continue to process the file and
ultimately -- in most cases -- correctly finish the upload.
Increasing the value should reduce the amount of confusion incurred for
users.