VidjilAuth.py add existance checks to permission checks
When calling a controller with an invalid ID, an Internal Server Error
will usually occur because the code was written with the assumption that
only valid ids would be passed to the controllers.

Not only can this be ambiguous for users since in this case there would
be no feedback from the server, but it may also give some information to
user with malicious intent.

With these changes, any permission check on an element which does not
exist will simply trigger the behaviour users would expect from
attempting to access existing data which they simply do not have
permissions for.