model_loader.js add id management to analysis samples
The bug described in issue #2176 , where an analysis may overwrite the
data of the wrong sample if the sample_set is altered after the analysis
has been saved, requires a sample identification method.
Therefore adding the original names to the analysis as a means of
identifying samples will achieve the end goal.