model.js: Set read number and top for added clone
The top is set to the highest possible value to ensure
that the clone will be visible.

The number of clones is set to a large arbitrary value.
May it happen that the clone doesn't appear because this
value would not be large enough?

Commit performed in trio-programming: @tydax @RyanHerb @mikael-s