generic_axis.js refactor label mapping
Changing the label storage to an associative array has several
advantages for storing and retrieving values within the axis.
However, not only did this change the external API, it also is not so
convenient for some tasks which were previously ver simple and intuitive
such as retrieving and iterating over the available labels.

Therefore having two label storage methods offers the best of both
worlds while reducing the overhead that recalculating a list of labels
would represent by mainting only the associative array.