bug20150629: Don't output a name if there is none.
This should be correctly handled in the browser but we make sure
that no name will be output if the name is empty.

This is the same behaviour as with -c segment.

grep was changed for sed because when no occurrence is found grep
exits with code 1, which is not the case of sed.