generic_axis.js, custom_axis.js pass values to constructor
We're moving towards compliance with the Liskov Principle. So all our
classes that inherit from GenericAxis need to be interchangeable
wherever possible.

CustomAxis used the data from the given instance of Model, but we want
to be able to theoretically apply an axis to any data. Therefore passing
the values to the constructor is a good way of making GenericAxis and
CustomAxis almost interchangeable.