segmentation-D-read-number.should-get: simplify the test
Why can't we keep complicated stuff when there exists really easy solutions?

The purpose is to test consistency of the segmentation with 1, 6 and 11 reads.
Instead of generating the data on the fly we can store the most abundant datasets
and use the magic -x option (see 7b314b614) to restrict the dataset to fewer
sequences.

This will solve some compatibility issues as this test could have some troubles
working on some architectures.