Makefile: Improve test speeds
Tests were launched with -O0 option to make debugging easier, in particular
with gdb but also possibly with Valgrind. However this induces a large time
penaly. As we want to improve test speeds (#2254), it should be a good idea to
use -O2.

Note that we do not compile with -DNDEBUG, which would improve speed test only
slightly. It is important that assertions are tested. Unit and functional
tests are the perfect place for that.