tests/models/VidjilAuth.py fix reversed tests
The test on can_modify_pre_process had inverted asserts. The user that
belonged to the admin group was expected to not be able to modify the
pre_process whereas the user that does not have that permission was
expected to have it.

Swapping the assert functions around resolves this issue.