SampleSet.py: Function to get the sample name
We would probably benefit from a class offering such functions.
The class would have a constructor relying only on the sample set ID
and would not need anything else.

The SampleSet class doesn't behaves in this fashion as this would not
be efficent enough (for the patient list for instance, see #2295).
However in some cases we would need a more flexible way of accessing
a sample set, even it is slower.