controllers/clonedb.py: Controller to call the CloneDB
Actually it is not a good idea that the client directly calls
the CloneDB. Indeed some tags from the CloneDB directly refers
to data on the server-side. Therefore the server can bring valuable
informations.

That's what we do know: the client calls a controller which calls
the CloneDB. The controller then adds other information useful to the client
(URL of the sample sets, name of the sample sets).