custom_axis.js refactor some steps_calculation
One of the operations on the calculations of steps got improperly
refactored when removing references to output. Hence we were
transforming nb_steps in ways that were designed to be used with data
such as floating points and "undefined" output as specified in previous
versions of the code.
This was breaking labels for percentage oriented data. Each class now
has its own method to properly treat the nb_steps variable according to
the data type.