generic_axis.js populateValueMapping on label_mapping keys
Not only is the code more elegant, but with the changes to the
functionning or the GermlineAxis, the allele axes bar charts were no
longer functionning.
Using the label_mapping keys rather than the label text produces the
same result for GenericAxis, but in the GermlineAxis, when using
alleles, the label_mapping keys for alleles contains the complete gene
name whereas the text only contains the allele reference.