clone.js fix variable reference
The variable 'field' was not defined in this context, but the issue was
not raised due to poor variable isolation in other parts of the code.

A global variable declaration was made in a loop in the model_loader.js file
causing the field variable to be accessible in the clone.js file, hence
the undefined variable issue was never raised with this piece of code.