numerical_axis.js fix identical labels
Some commits seem to have been lost or forgotten at some point in
regards to the positionning of the 'undefined' (?) values label.
And we had identical labels for all positions on numerical axes. The
position of the 'undefined' label was also reset at some point as well
as the label offset.

This should sort things outs and reposition all labels correctly with
the correct text.