browser/tests/*.rb: Fix tests for IMGT productivity
Due to the changes in axes management, there was some changes in the way stuff is displayed
in the HTML page.
Therefore the clone_info_segmenter also has to return the axisBox span (which didn't exist
before and whose content was in infoBox).
The productivity information can now be textually found in the title of the
span.

See #2412