docker-compose.yml link conf files to host.
In an attempt to make it easier to alter the configuration of the
dockerised Vidjil installation, the host now contains the cofiguration
files within a volume linked to the containers.
This modification should allow users to modify some configurations while
the app is running, and should the containers require restarting, the
image no longer needs to be rebuilts for the changes to take effect.