file.py fix sample_set_membership deletion
The loop in charge of deleting sample_set_membership was placed after
the deletion of the sequence_file, which caused the references to said
sequence_file to be set to NULL, and therfore the loop was never
executed.
Therefore placing the loop ahead of the sequence_file deletion should
reduce database pollution.