file.py, task.py refactor schedule_fuse
The previous implementation of schedule fuse would only refuse a single
sample_set and a single config. The goal of the function being to refuse
all fused_files associated with a removed sequence_file, this was
obviously not adequate.
The function now takes a list of sample_set ids and a list of config ids
in order to detect at one sequence file per sample_set and per config
and refuse these results.