VidjilAuth.py refacotr permissions cache.
Changeing the was cached permissions are stored into the form
cache[table][record_id][permission] allows for easier verification of an
objects presence in the cache.
This will be useful for avoiding queries to the database when checking
if an object exists, since it will be easy to see if it has been loaded into the cache.