algo/: Refactor Fasta to BioReader
We now have an abstract class to deal with biological sequence files. This
will allow to more easily manage different file types.

This commit only reorganizes the code so that we will be able to add a BAM
reader easily. Functionnally the code should be equivalent to its previous
version.

Some functions that were not used have been removed.

The operator>> has been removed as it was only used in unit testing. This
operator is not convenient as having the filename may be useful to reopen the
file or to know its extension, to guess the filetype.

See #2016