Replace jquery.autocomplete.js with jquery.atwho.js
Although the jquery plugin autocomplete.js was providing the basic
functionnality we needed, it was not natively able to display the list
of suggestions in the desired position for our textareas.

At.js is capable through the use of Caret.js of displaying a suggestion
box under the caret in a textarea.

This javascript code is heavily inspired by the code present in the
gitlab-ce project and allows us to have a nicely cached data-set to
limit database queries.