nginx_web2py, nginx_install.sh change cache control
Cached items are causing issues after updates, since the server
sometimes expects new elements in requests which the client-side files
are not yet able to provide (due to caching conserving older versions of
ressources).

Altering the caching policy to must-revalidate forces the client's
browser to check with our server whether a file has changed, allowing
nginx to respond with code "304 Not Modified" when appropriate, and
serving the new versions of outdated ressources correctly.

See #2498