automaton.hpp: Count kmers even if they are not new
This seem weird and is a bit hacky. But the purpose is
not to disadvantage a real germline (eg. IGH) compared
to a fake germline (unexpected).

As there are more ambiguous k-mers in the unexpected germline
there will be less IGH k-mers hence a smaller index load for
IGH in the unexpected germline rather than the real one.

See also #2651.