sample_set_list.py, models/tag.py case insensitive first steps
In order to keep things simple for the time being, searching and saving
tags is now case insensitive.

Example: registering #Tag when #tag already exists will will simply link
the element to the existing #tag.
Searching for #Tag will also return #tag.

The downside to this is that auto_completion will always suggest what
the first user registered as opposed to that which the user is trying to
type.

See #2671