file.py, file forms, database.js start file autocompletion from scratch
This is to start addressing a two issues that we are having with file
uploads.
The first being maintainability. Having three fields that perform the
same task with three times the same code isn't acceptable. This way, the
three fields are generated by the same code.
The second, is an issue with datalists. They don't seem to work
consistently accross all browsers, especially in older versions. And
they don't accept css, or provide a scrolling feature.

The datalists have now been removed, helping to make the controller a
little faster.