packaging+docker package web2py source code with vidjil-server
The web2py package for debian is a little out-dated and we would like to
use the latest version. As mentionned here web2py.com/init/default/download
specifying in the license where web2py's code begins and our app's ends
allows us to do this.

This commit updates the copyright file in the package to provide these
details and allow us to package our server with whichever version of
web2py we need.