model.js model_loader.js move applyUrlParams to init()
The point being, that the scatterPlot requires some functionnality
called by initClones(), however, initClones() ends with a call to
init(), which leads to a call to update(), which would override the
selected clones parameter in the URL (calling telling Url.js to update
the url before applying the url params), which is why the call the
applyUrlParams() was placed before initClones() in the first place.

So the solution at the time being seems to be to call applyUrlParams()
in the init function, before it leads to an update().