scatterPlot.js undo commented changePreset
A changePreset in the scatterPlot init method had been commented during
development of this feature, due to some issues it was causing.
Improvements in the implementation of this feature mean that function is
no longer an issue, and in fact is necessary for the multiple
scatterPlot view to function correctly.