scatterPlot.js new methods for updating plot settings
updatePreset and updateMode are here to simply mark off a user action of
changePreset and changeMode respectively. Since changePreset (and
potentially changeMode) could be called within code, we don't
necessarily want them to trigger an update. But in the case of
updatePreset and updateMode, we want an update to be triggered for
the url params to be updated correctly.