database.js fix upload forms
Recent changes to the flow of the client, namely the removal of several
calls to model.initClones, seem to have either caused, or brought to
light another application flow issue.
Where, when the server was returning an indication to the application
that it should redirect the user to a new view, the call to the new view
was called, but the execution of database.display_result(result) was
continued, therefore attempting to identify further formats of potential
server responses. Given that in such a context, there may be no samples
initialised in the model, the application was crashing, therefore not
cotinuing it's execution of the upload form controller initialisation.

This was causing the uploads to never occur.
See #2849