tools-vdj-assign-direct-input.should-get: Use 'normal' sequence
The former test was done on a sequence which was V only.
The purpose of this test should be that the behaviour of
the program is the same as when providing the sequence
in parameter. Therefore the sequence used is now the same
as in tools-vdj-assign.should-get, and the expected output
is also the same.