should-vdj-to-tap.py: Count the number of bugged assignations
Each bugged sequence must be tagged correspondingly. However the benefit is that
we don't need to update a counter without knowing what sequence has changed.

There is still a (minor) drawback: only the count is actually checked. Therefore
we could have some newly failed sequences balancing sequences not failing anymore.
However in the output we will see the failed sequences that should have not.

This is another benefit: the output is now much clearer as we only see sequences
that have failed (and should have not) or sequences that succeeded (and should have not).