0000-nck-TRB.should-vdj.fa: Bad assignation relying on homopolymer
The former assignation was assuming the following alignment at the end of the V:
GCTTTAGC
|||| |||
GCTT-AGC
by considering that the TTT aligns with TT as it is an homopolymer.

This could be, however since those sequence comes from an Illumina, it is
highly unlikely.

Even when forbidding homopolymers this sequence is still badly assigned. The
reason why is unknown (see #2904).