segment.h: The minimal window length was actually quite small.
It was hard to make some tests (eg. unsegmentation-causes.should-get) have a
too short window. Modifying a constant just so that the test passes is bad
practice, but this is not the case here (at least I believe so).  Having
difficulties to have a too short w also meant that in practice we would almost
never see a “too short w” again. This could be nice. But we know that the
position of the window is not accurate (it is not designed to be). Therefore
20nt, if they are perfectly centered may be ok. But we know that it may not be
well centered (and we even allow some shifts). Therefore 30nt seems to be a
safer limit.

See #2913