core/{windowExtractor,windows,segment}: Short/shifted windows
Warn when we have one.

This requires to add a new segmentation status. The problem is that
Segment::getSegmentationStatus() returns just an integer which doesn't make
possible to have several statuses at once (which could be convenient). A
solution would be to use masks for the statuses (ie. if s is the status we
would store 1 << s).  Since this would require quite some code modification to
achieve that, I added some extra tests (that's bad) to deal with the case
where the segmentation is achieved with a shorter or shifted window. In such a
case we warn the user in the JSON (warning on the CLI should be done too).

See #2913, #2916.