migrator.py refactor id replacement method
A more flexible and explicit method of selecting the target field for
foreign_key replacement.
Appending "_id" to the table name (stored as the key for the mapping)
does not work for al situations, especially if managing table that were
not created for the current application, so explicitly mapping the
mapping key to the target field will be more flexible.