autocomplete.js change matcher for samples.
Due to the lack of a flag to start matching text, it was necessary to
alter the matcher to allow us to start matching the last element in an
input without. The lack of a flag means that all the text in the input
is matched to the query and therefore we needed to extract the last
matching string to autocomplete.