file.py failsafe if initial set token is removed
The redirection at the end of the submission controller checks the
initial sample-set-type in order to redirect to the source sample-set.
But if the initial sample-set is removed from the list of sets to add a
sample to, then we need to ensure we still have a set to redirect to.