refactor patient add/edit into a single controller and form
Having a single controller and view, is better for clean code practices,
since changes to the form/controller won't require two forms and
controllers to be changed.

So when the changes for the multi-patient form come in, we'll have less
to do and a cleaner code-bade to work from.

I also took the opportunity to build the from with divs rather than a
table since we're going to need a more flexible layout when the time
comes. Also less DOM nodes.