Refactor file controller and view
split the actual form part into a partial as with sample_sets, this
allows us to have shorter files with less levels of indentation, making
a little easier to read.

also applies the multi-form naming convention so we can apply
serializeObject to the form.

refactors a couple elements from the controller into functions (such as
validation and pre_process listing) to shorten the controller itself.