jstree onload
For clarity purposes we don't want the users to have to trigger the
mouseover event in order to display the jstree file selector. So using
the onload event is much better in that respect. Howver given the way
these DOM elements are loaded, that event is not triggered by the
browser. Therefore we need to trigger it ourselves.

For this we need to target the specific DOM node we want the event to be
triggered on since the 'load' event does not bubble, so triggering it on
for instance the div with id "db_content" does not work for us.