refactor file form add button
There was too much happening in the onclick to keep it in the HTML. On
top of needing to apply a selector with characters that re forbidden in
HTML attributes.

This refactor was necessary in order to check the value of the radio
buttons that determine the file source. This way we can display the
correct fields when adding a file form not matter which file source is
currently selected.