refactor Tokeniser delegate to autocomplete
tokens are now created from text rather than by reading the value of an
input. This allowed us to delegate the token creation to the
autocomplete beforeInsert callback for Samples.
This not only makes it impossible to create "trash" tokens. Tokens can
only be created as a result of an autocompletion, but it also means that
we are not binding events to the input and any selection method of an
autocomplete suggestion will create the token