testRunner.py clone fields on db copy
what seems to be some recent changes to pyDAL are causing our method for
copying fields to fail since we are trying to assign a field to a new
table while it is already assigned to it's parent. Cloning the field
allows us to have a new field assigned to the new table.