database.js extract args and next from login action
due to some issues with the redirection of the default web2py login controller, the ajax callback expected to be called was the error callback and managed the url parameters correctly. However, this controller seems to be functionning correctly now and therefore args were not correctly passed on to the following page, causing an issue when attempting to load an analysis from the login form (logging in after being redirected for attemting to load an analysis while not logged in).

Now we just manage the args in the same way whether the controller fails
or succeeds.

See #3072