database.js db_layout.html fix logout redirection
Much in the same way the login form started functionning, the correct
procedure was not being applied once the logout form started "working".

So we now created a separate method for logout since it was safer the
request method was not appropriate for the current behaviour of the
logout controller.

See #2990