views/{sample_set,file}/form.html: Bikeshedding on 'add' buttons placement
It should be better to have the add buttons before the form rather than after.
When they are after they move each time a row is added. Therefore if one knows
that s⋅he needs 10 samples, s⋅he couldn't click 10 times quickly on the link,
s⋅he had to click, wait for the new row to appear, adjust the mouse position,
click, and so on…