vidjil_utils.py: Seems that safe_encoding is no more needed now
See #3097

Discussed with @magiraud and @RyanHerb: The UTF-8 is now properly handled when
processing JSON. An example of that is that runs and sets don't need safe_encoding.

However there is still a small issue (that pre-existed): in anon mode, the three
first bytes are recovered which may not correspond to a valid character.