js/autocomplete.js: Make sample set autocompletion much faster
The RegExp used was taking ages in some cases its purpose was to put in bold
the query when it was not in the enclosing tags.

We modified that with @RyanHerb to first isolate the tags
and then replace all the occurrences of the query to put them
in bold. A RegExp is still useful (we need to find all occurrences).

This fixes partially #3100 (another part of the problem is on the server side)