controllers/file.py: Change group IDs for tag autocompletion
After a pair-code-review with @magiraud this fixes #3121

The tag autocompletion was not the same for the search box on the sample set
page and when adding a sample or a set. This was due to the group IDs passed
to the autocompletion that didn't include the public group. Therefore we (with
@magiraud) chose to take the same code that worked fine in sample_set.py for
defining group_ids.