data/segmentation-2.fa: Small change in random sequence
That random sequence was supposed to have too few v/j information (because
it's random!). However, because of bad luck there was some tiny similarities
with IGH but the sequence was not segmented due to inconsistencies with the
strand.
This is still cool because the random sequence is not segmented. However here
we just want to test that a sequence with too few v/j is reported as such. As
the strand consistency is tested first in our algorithm, we didn't report the
too few v/j information. Therefore I modified a single nucleotide in the
sequence to get rid of this strand inconsistency.

I think I always feel guilty when modifying a test example. That's probably
why I needed to make such a long comment.

See !148