segment.cpp: Estimate number of aligned nucleotides
Knowing the number of aligned nucleotides allow to compute the minimal score under
which the optimisation may have biased the result.

The number of aligned nucleotides could be computed more accurately with #2138.
Indeed if there are large portions that are deleted in the read (for instance)
they won't be taken in the number in this case. However it would be important
as the optimisation may yield a wrong gene in such a case.

See #3066