docker* add test build environment
allows the use of the 'build_env' build arg which will build the docker
image in "test" mode.
This has two consequences. The first is that all ocnfigurations files
are set to run the env over HTTP for ease of use. The second is the
installation of the necessary packages and programs required to run
watir functional tests in headless mode.

See #2881