Get the ex-aequo best k-mers
If two K-mers occurs both the same amount of times, we want them both. Even if
the number of significant K-mers is reached.

For example if we want the 10 most significant K-mers but the 4th has two K-mers
with the same value we take them both instead of one.

For more informations, see #3217