change comparator while sorting map in filterBioReader
The comparator checks if the two KmerAffect have the same number of
occurences.
If it is the case, then the Comparator use the comparison operators defined
in the KmerAffect class to compare the KmerAffects.
If the two KmerAffect doesn't have the same number of occurences, the Comparator
returns true for the one that have the most of them.

For more informations see #3217