simplify and fix gene counter while filtering
The gene counter was already at 1 when doing the first loop since the
previousKmerOccurs was initiated with '0'. To avoid that, it's now initiate
with the value of the first member of the set.

Also a simple simplification for the gene counter (deleted the empty if and
replaced it with the opposite condition)

For more informations #see 3217