vidjilparser.py fix comma delimitation on output
The introduction of the buffer caused an issue with comma delimitation
when th first element of an array or map was discarded (from not meeting
predicate requirements), leading to an extra comma and therefore an
invalid JSON output.
By saving the value of the "previous" variable when buffering begins, we
can restore the variables value when discarding the buffered content.

See #3235