add -Z in getopt and get the value
In #3268 we want to add a specific command-line option for the filtering.
The -Z has only been declared in the getopt. The remaining task is to
check if -Z still has NO_LIMIT_VALUE as an argument. If so the automaton
is not built, otherwise the automaton is built and the argument is used to
filter the BioReader.