testAffectAnalyzer: Test two cases where we have overlapping before/after affects
This causes the heuristic not to be symmetric because and having
an undesired behaviour.
In such a case we have a plateau with the maximal value. This plateau
is reached when we have the overlapping V/J. But in this case the J
affects are counted « left » which then prevents the heuristic from
segmenting the sequence as the number of after affects (the Js)
on the left part is too high.

Therefore this has to be changed to that the overlapping affects
are not counted anymore as left or right.

See #3296