testSegment.cpp: Not ambiguous anymore
The test was classified as ambigious but it probably should not have been. The
fact of having not enough Vs should have classified the read as UNSEG_ONLY_J.

Now the overlapping k-mers where we maximize the score (with the 'before'
affects at the left and the 'after' affects at the right) are ignored. This
therefore gives us the UNSEG_ONLY_J.

The previous UNSEG_AMBIGUOUS was due to the fact that some 'after' affects
(the Js overlapping the end of Vs) were counted left affects and that there
was too many left Js.

Thus we now have 4 UNSEG_ONLY_J sequences and 1 UNSEG_AMBIGUOUS.  As the
sequence whose unsegmentation cause changes is 33nt long, the average length
changes accordingly.

See #3296