core/affectanalyser.cpp: Don't count affects that are on a plateau of max values
When they were counted this biased the counter of before affects on the left part thus
leading to unjustified UNSEG_AMBIGUOUS.

When we are on a plateau and when affectations[i] is defined this means we have overlapping k-mers
thus we don't count those k-mers as left affects (are they left or right? we are actually **on**
the maximum, it is hard to tell).

However if we happen to encounter a higher maximum those ignored values are then reaffected to the left
counters.

Fix #3296