Fix testGetMultiResults when there is only unknown K-mer
The previous test was only looking for a single unknown K-mer in the map even
though the map shouldn't contain anything else from unknown K-mers.
Indeed when there is no K-mers in the automaton that match the input sequence,
then the map contains only unknown K-mers.
For more informations, see #3303.