tools/create-git-sha1.sh: Don't touch the git file.
Touching it updated its last modification date which triggered the recompilation of vidjil.cpp.
We don't touch it anymore. Therefore the file may not exist and the diff may fail. We just ignore
error messages. In such a case the mv will be executed and the file will then exist.

Fixes #3309