algo/Makefile: Remove base directory with lcov
lcov needs to know the whole path to a file in order to display it
and to count the coverage on each file.

However when launching with Gitlab-CI, each job is launched in a separate directory
which leads to different paths for the same source files.
Therefore we first remove the trailing path when generating the
lcov files. But we add it again before generating the HTML output
as it requires to access the files.