Add isGeneric and SPECIFIC_KMERS_NUMBERS in kmerAffect
isGeneric is a method returning true if the current K-mer is not odd (ambiguous, unknown).
It is used instead of calling successively isAmbiguous() || isUnknown()...

SPECIFIC_KMERS_NUMBER is a constant representing the amount of odd K-mers. For now there is
only two (ambiguous/unknown) but in the future it might grow.

For more informations about it, see #3282.