add test for transferBioReaderSequences();
The test verifies that when an odd K-mer is transmitted (unknown, ambiguous, wrong ascii n° for its label...)
an exception is thrown. Also it verifies that when a K-mer is transmitted, there is the good amount of sequences
that are transfered from a BioReader to another.

For more informations see #3282.