algo/vidjil.cpp: enable filtering by defaut, -Z 3
Thanks to the work by @boreec, we now enable by default
the filtering for the second pass of the clone analysis.

When the computing time is dominated by the FineSegmenter
(such as with '-c clones -z 10000' or '-c segment'),
we get final speed-ups between 2x and 15x (#2360).

We could even set -Z 1 in most of the cases, but it should
be coupled with #3225. See also #3297 and #3344.

Close #3298.