segment.{h,cpp}: Make sure number of matches is > 0
When computing whether the alignment has too many differences
we should not allow having a negative number of mismatches.
So we have to make sure that length >= BOTTOM_TRIANGLE_SHIFT.
But this is not enough: having only 1 match in 20 nucleotides
doesn't make sense. Therefore we may allow a given fraction
of the sequences to be matches just by chance.
We could have used probabilities but this should be a rare
case. I'm not sure that's worth it.