test_sample_set.rb more robust edit test
Since the order of tests is no longer garranteed, we can't expect the
index of patients to be the same each time. So we need to select our
patient based on another field (like info). And use the sample_set_id to
retrieve it after the info has been edited.