js/{generic,numerical}_axis.js, js/scatterPlot.js: rename and set MAX_NB_STEPS_IN_AXIS to 8, but stay at 6 for Y
As the mechanism is more generic that what was used by bb410296, we can now have
a higher maximal number of steps, at least for X.
The actual numbers of steps will be often smaller, adapted to a nice distribution of label values.

Closes #2699.

One day, these values will be fixed dynamically to be responsive. See #2450.