sample_set.py, stats_decorator.py refactor query and set info
We don't want files to appear multiple times if they are present in
multiple sample sets. In order to achieve this, we retrieve the
information from the results_file and group all the sample_sets into an
array for each results_file. A new SetDecorator displays the list of
sets how we want them.
See 3407