algo/core/filter: Don't have a hard limit on number of sequences.
Rather we use an uncertainty range (see #3225 for why and how).

It is necessary to get the length of the longest transferred sequence as it
will be used as a reference to compute the empirical probablity that a k-mer
has matched in the longest possible sequence. Therefore we make the assumption
that k-mers matched in the longest sequence but we have no way of knowing as
we don't keep track of where the k-mers matched.

Fixes #3225