HACK -- algo/core/germline: Change shortcut to avoid ambiguity
There could have ambiguity: for instance no way of distinguishing
labels coming from IGH and IGH+. While it could make sens for the J
genes as they are the same it was more annoying for the 5' gene as
they differ in both loci.

This is somewhat solved but there are limitations.
We cannot easily solve the case where some of the 5' (for instance)

See #3584