bug20160121.should-get: Change test e-value
The test passed some times ago (see 8456dbad) probably due to a slight change
in the e-value. Now with Aho-Corasick we have slightly different e-values.

It would be a pity to put this test as failing. I prefer lowering the evalue
threshold to make sure that we don't have any regression in the future.