vidjil.less list.js fix css remove need for resize
Since we are no longer using table-cell to display the list, we can do
away with the resize code which was designed to ensure the scrollbar get
added correctly to the div.
With the flex styling we just need to set an arbitrary height to 0px,
which will be overridden by the flex styling.