Update Watir
Use latest Ruby version and Watir instead of watir-webdriver.
This will make it easier to install it on recent OS and it
will also allow us to test on recent Firefox versions.

Some changes in the syntax:
- visible? is deprecated → present?
- wait_until_* are deprecated → wait_until(&: *)
- some way of specifying timeouts

Watir also seems to be quicker thus the need to add
some waits or some sleeps