algo/tests: Test the --consensus-on-random-sample option
The dataset was generated to have a minority of sequences with a TTT insertion
which would make it the sequence of choice with the default ReadScore (based
on length and quality).

On the contrary with a random sampler we should not be impacted by this
insertion as it happens in the minority of sequences.

Here is the command that generated the dataset:
for i in $(seq 1 3500); do
  if [ $((RANDOM%3)) -ne 0 ]; then
    echo ">seq$i";
    echo ctacctactactgtgccttgtgggaggtgatagtagtgattggatcaag;
  else
    echo ">seq$i";
    echo cTTTtacctactactgtgccttgtgggaggtgatagtagtgattggatcaag;
  fi;
done > test-random-consensus.fa