vidjil.cpp: Fix definition of min_cover_representative
min_cover_representative should be seen as a strict minimum on the number of reads
that can support a representative sequence.
The purpose is to avoid a too low ABSOLUTE number of reads to define a representative.
For the relative number, it is ratio_representative's role.

Before, if we specified -r 100 we needed to have 100 corresponding
k-mers which is very strict when a clone has just slightly more than
100 reads