index.html + css: display right part of the menu with flex
This requires to reorder elements in the HTML.
Thus they are now in a 'natural' order (ie. in the same order
as displayed). We use flex to make them appear contiguously the
ones after the other. Thus we don't need anymore float: right
apart from the #menu_right_part that was added.

Fix #3805