server/.../models/{db,task}.py: Results can be hidden
We add a new field to the results_file stating that they can be hidden.
This means that those results shouldn't appear in the last processed
results. They should not be taken into consideration when fusing results.
They should not be proposed when comparing results.

A first example of those hidden results is when we launch a grep reads.
This is just a process launched behind the scene that should not be
made explicit to the user.

Fix #3899