views/sample_set/index.html: Allow to launch process depending on pre-process
We don't want to allow launching an analysis while one of the file is still uploading (which is the
case with a pre-process with two files). Thus to be able to launch a process we expect either
to have no pre-process or to have one but already finished. We allow to launch a process even
if the pre-process failed to prevent the user from being blocked (he/she can still analyse
the R1 file).