testAffectAnalyser: Don't pass the empty string in parameter
As the parameter is a reference, it may cause some problems if it is
optimized out at some point (or something else, I won't pretend mastering
compiler optimizations). Instead we pass a variable in parameter and instead
of an empty string we give a string whose length is the same as the affectation
string.

This was causing issues with clang-4.0 and clang-6.0.

See !612