automaton.hpp: Copy information when defining new defining new final states
When finalising the automaton we can define new final states that
actually are a “copy” of another final state. In such a case we
also need to copy the Info that are in that former final state.

Added a test to check that.
The test uses a IGKJ sequence: thus all the kmers should be +k
or (potentially undetermined).

Fix #4225